+++ src/data/api/__mocks__/SpotifyApi.ts
@@ -1,11 +1,10 @@
 import { AxiosResponse } from "../../../../node_modules/axios";
 import { resolve } from "path";
+import { SpotifyApi } from "../SpotifyApi";
 
-export class SpotifyApi {
+export class MockSpotifyApi implements SpotifyApi {
 
     constructor() { }
-
-    public getAccessTokenRequest =  jest.fn(() => { });
 
     public getAlbums = jest.fn((album: string, authorization: string) => {});
 }
