--- src/domain/repository/AuthRepository.ts
@@ -1,3 +1,0 @@
-export interface AuthRepository {
-    getAccessToken(needRefresh: boolean): Promise<string>
-}
