--- src/domain/repository/__mocks__/AuthRepository.ts
@@ -1,5 +1,0 @@
-import { AuthRepository } from "../AuthRepository";
-
-export class MockAuthRepository implements AuthRepository {
-    public getAccessToken = jest.fn(() => { });
-}
